+++ src/main/java/com/sumologic/jenkins/jenkinssumologicplugin/sender/SumoPeriodicPublisher.java
@@ -9,27 +9,21 @@ import com.sumologic.jenkins.jenkinssumologicplugin.model.QueueModel;
 import com.sumologic.jenkins.jenkinssumologicplugin.model.SlaveModel;
 import com.sumologic.jenkins.jenkinssumologicplugin.utility.CommonModelFactory;
 import hudson.Extension;
-import hudson.Util;
-import hudson.model.*;
 import hudson.model.Queue;
+import hudson.model.*;
 import hudson.model.queue.WorkUnit;
-import hudson.node_monitors.NodeMonitor;
 import jenkins.model.Jenkins;
 import org.apache.commons.collections.CollectionUtils;
 
 import java.io.IOException;
-import java.lang.reflect.Method;
 import java.util.*;
 import java.util.concurrent.TimeUnit;
 import java.util.logging.Level;
 import java.util.logging.Logger;
-import java.util.regex.Matcher;
-import java.util.regex.Pattern;
 import java.util.stream.Collectors;
 
-import static com.sumologic.jenkins.jenkinssumologicplugin.constants.SumoConstants.*;
+import static com.sumologic.jenkins.jenkinssumologicplugin.constants.SumoConstants.DATETIME_FORMATTER;
 import static com.sumologic.jenkins.jenkinssumologicplugin.utility.CommonModelFactory.*;
-import static org.apache.commons.lang.reflect.MethodUtils.getAccessibleMethod;
 
 /**
  * Created by deven on 8/6/15.
@@ -42,12 +36,13 @@ public class SumoPeriodicPublisher extends AsyncPeriodicWork {
     private static final long recurrencePeriod = TimeUnit.MINUTES.toMillis(3);
     private static final Logger LOGGER = Logger.getLogger(SumoPeriodicPublisher.class.getName());
     private LogSender logSender;
-    private Set<String> slaveNames = new HashSet<>();
-    private static LogSenderHelper logSenderHelper = LogSenderHelper.getInstance();
+    private LogSenderHelper logSenderHelper;
+    private static Set<String> slaveNames = new HashSet<>();
 
     public SumoPeriodicPublisher() {
         super("Sumo Logic Periodic Data Publisher");
         logSender = LogSender.getInstance();
+        logSenderHelper = LogSenderHelper.getInstance();
         LOGGER.log(Level.FINE, "Sumo Logic status publishing period is {0}ms", recurrencePeriod);
     }
 
@@ -72,7 +67,7 @@ public class SumoPeriodicPublisher extends AsyncPeriodicWork {
         return recurrencePeriod;
     }
 
-    private void sendTasksInQueue() {
+    public void sendTasksInQueue() {
         final Queue.Item[] items = Jenkins.get().getQueue().getItems();
         List<String> queueModels = new ArrayList<>();
         for (Queue.Item item : items) {
@@ -95,9 +90,8 @@ public class SumoPeriodicPublisher extends AsyncPeriodicWork {
         logSenderHelper.sendMultipleLogsToPeriodicSourceCategory(queueModels);
     }
 
-    private void sendNodeDetailsForJenkins() {
+    public void sendNodeDetailsForJenkins() {
         List<SlaveModel> slaveModels = getNodeMonitorsDetails();
-        LOGGER.info("slave models details "+slaveModels.size());
         if (CollectionUtils.isNotEmpty(slaveModels)) {
             List<String> messages = new ArrayList<>();
             slaveModels.forEach(slaveModel -> messages.add(slaveModel.toString()));
@@ -123,98 +117,7 @@ public class SumoPeriodicPublisher extends AsyncPeriodicWork {
         slaveNames = slavesUp;
     }
 
-    private List<SlaveModel> getNodeMonitorsDetails() {
-        List<SlaveModel> slaveModels = new ArrayList<>();
-        Computer[] computers = Jenkins.get().getComputers();
-
-        if (computers == null || computers.length == 0) {
-            return slaveModels;
-        }
-        Collection<NodeMonitor> monitors = ComputerSet.getMonitors();
-        for (Computer computer : computers) {
-            if (computer != null) {
-                SlaveModel slaveModel = new SlaveModel();
-                slaveModel.setLogType(LogTypeEnum.SLAVE_EVENT.getValue());
-                slaveModel.setEventTime(DATETIME_FORMATTER.format(new Date()));
-                slaveModel.setEventSource(EventSourceEnum.PERIODIC_UPDATE.getValue());
-                getComputerStatus(computer, slaveModel);
-                for (NodeMonitor monitor : monitors) {
-                    slaveModel.getMonitorData().putAll(getMonitorData(computer, monitor));
-                }
-                slaveModels.add(slaveModel);
-            }
-        }
-        return slaveModels;
-    }
-
-    public static void getComputerStatus(Computer computer, SlaveModel slaveModel) {
-        slaveModel.setNodeName(getNodeName(computer));
-        Node slaveNode = computer.getNode();
-        if (slaveNode != null) {
-            slaveModel.setNodeLabel(slaveNode.getLabelString());
-        }
-        slaveModel.setNodeStatus("updated");
-        slaveModel.setNumberOfExecutors(computer.getNumExecutors());
-        slaveModel.setIdle(computer.isIdle());
-        slaveModel.setOnline(computer.isOnline());
-        if (computer.isOffline()) {
-            slaveModel.setNumberOfExecutors(0);
-            slaveModel.setRemoved(true);
-            slaveModel.setReasonOffline(computer.getOfflineCauseReason());
-            slaveModel.setConnecting(computer.isConnecting());
-        }
-        slaveModel.setNodeURL(getAbsoluteUrl(computer));
-        long connectTime = computer.getConnectTime();
-        if (connectTime != 0) {
-            slaveModel.setConnectTime(Util.XS_DATETIME_FORMATTER.format(connectTime));
-        }
-    }
-
-    private static String getAbsoluteUrl(Computer computer) {
-        String rootUrl = Jenkins.get().getRootUrl();
-        if (rootUrl == null) {
-            return computer.getUrl();
-        } else {
-            return Util.encode(rootUrl + computer.getUrl());
-        }
-    }
-
-    private static String getNodeName(Computer computer) {
-        if (computer instanceof Jenkins.MasterComputer) {
-            return MASTER;
-        } else {
-            return computer.getName();
-        }
-    }
-
-    private Map<String, Object> getMonitorData(Computer computer, NodeMonitor monitor) {
-        Map<String, Object> monitorDetails = new HashMap<>();
-        Object data = monitor.data(computer);
-        if (data != null) {
-            String monitorName = monitor.getClass().getSimpleName();
-            String monitorData;
-            Method method = getAccessibleMethod(data.getClass(), "toHtml", new Class<?>[0]);
-            if (method != null) {
-                try {
-                    monitorData = (String) method.invoke(data, new Object[0]);
-                } catch (Exception e) {
-                    monitorData = data.toString();
-                }
-            } else {
-                monitorData = data.toString();
-            }
-            Pattern compile = Pattern.compile(ERROR_SPAN_CONTENT, Pattern.CASE_INSENSITIVE);
-            Matcher matcher = compile.matcher(monitorData);
-            if (matcher.find()) {
-                monitorDetails.put(monitorName, "warning:" + matcher.group(1));
-            } else {
-                monitorDetails.put(monitorName, monitorData);
-            }
-        }
-        return monitorDetails;
-    }
-
-    private static void sendRunningJobDetails() {
+    public void sendRunningJobDetails() {
         List<String> currentBuildDetails = new ArrayList<>();
         for (Computer computer : Jenkins.get().getComputers()) {
             List<Run> runList = new ArrayList<>();
